requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dicom2jpg==0.1.10
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
gevent==24.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.4.1
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
psycopg2==2.9.9
pydicom==2.4.4
Pygments==2.18.0
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pylibjpeg-openjpeg==2.3.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
rarfile==4.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
unrar==0.4
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
zope.event==5.0
zope.interface==6.4.post2
